+++ app/Filament/Resources/Articles/ArticleResource.php
@@ -19,6 +19,7 @@ use App\Filament\Resources\Articles\Pages\ListWords;
 use App\Filament\Resources\Articles\Pages\ViewWord;
 use App\Filament\Resources\Articles\RelationManagers\LabelsRelationManager;
 use App\Filament\Resources\Articles\RelationManagers\NotesRelationManager;
+use App\Filament\Resources\Articles\Schema\ArticleForm;
 use App\Filament\Resources\Articles\Schema\FormSchema;
 use App\Filament\Resources\Articles\Schema\WordInfolist;
 use App\Filament\Support\Concerns\HasActiveIcon;
@@ -160,205 +161,7 @@ final class ArticleResource extends Resource
      */
     public static function form(Schema $schema): Schema
     {
-        return $schema->components([
-            Group::make()
-                ->schema([
-                    Section::make('general-information')
-                        ->compact()
-                        ->collapsible()
-                        ->collapsed()
-                        ->columns(12)
-                        ->heading('Algemene informatie')
-                        ->icon(Heroicon::OutlinedInformationCircle)
-                        ->iconColor('primary')
-                        ->description('De algemene informatie van het artikel in het Vlaams woordenboek')
-                        ->schema([
-                            Select::make('partOfSpeech')
-                                ->label('Woordsoort')
-                                ->columnSpan(3)
-                                ->relationship(titleAttribute: 'name')
-                                ->optionsLimit(4)
-                                ->searchable()
-                                ->preload(),
-                            TextInput::make('word')
-                                ->label('Woord')
-                                ->columnSpan(3)
-                                ->required()
-                                ->maxLength(255),
-                            TextInput::make('characteristics')
-                                ->label('Kenmerken')
-                                ->columnSpan(6)
-                                ->required()
-                                ->maxLength(255)
-                                ->helperText('Dit veld is verplicht. Maar als er geen kenmerken zijn, vul dan \'-\' in.'),
-                            TextInput::make('keywords')
-                                ->label('Kernwoorden')
-                                ->translateLabel()
-                                ->placeholder('Kernwoord 1, Kernwoord 2, Kernwoord 3, etc...')
-                                ->columnSpanFull(),
-                            Select::make('labels')
-                                ->relationship(titleAttribute: 'name')
-                                ->multiple()
-                                ->preload()
-                                ->native(false)
-                                ->columnSpanFull(),
-                            TextInput::make('image_url')
-                                ->label('Afbeelding')
-                                ->columnSpan(6)
-                                ->url()
-                                ->prefixIcon('heroicon-m-globe-alt')
-                                ->prefixIconColor('primary')
-                                ->helperText(str('**Gelieve enkel afbeeldingen van wikipedia te gebruiken**')->inlineMarkdown()->toHtmlString())
-                                ->maxLength(255),
-                            TextInput::make('image_alt')
-                                ->label('Afbeelding alt tekst')
-                                ->columnSpan(6)
-                                ->maxLength(255)
-                                ->placeholder('Beschrijf kort wat er op de afbeelding staat')
-                                ->prefixIcon('heroicon-m-chat-bubble-bottom-center-text')
-                                ->prefixIconColor('primary'),
-                            MarkdownEditor::make('description')
-                                ->label('Beschrijving')
-                                ->columnSpanFull()
-                                ->toolbarButtons(self::getToolbarOptions())
-                                ->placeholder('De beschrijving van het woord dat je wenst toe te voegen.')
-                                ->helperText(str('Dit veld ondersteunt enkel [**Markdown**](https://www.markdownguide.org/cheat-sheet/)')->inlineMarkdown()->toHtmlString())
-                                ->maxHeight('125px')
-                                ->required(),
-                            MarkdownEditor::make('example')
-                                ->label('Voorbeeld')
-                                ->toolbarButtons(self::getToolbarOptions())
-                                ->placeholder('Probeer zo helder mogelijk te zijn')
-                                ->helperText(str('Dit veld ondersteunt enkel [**Markdown**](https://www.markdownguide.org/cheat-sheet/)')->inlineMarkdown()->toHtmlString())
-                                ->columnSpanFull()
-                                ->maxHeight('125px')
-                                ->required(),
-                        ]),
-
-                    Section::make('status-information')
-                        ->compact()
-                        ->collapsed()
-                        ->collapsible()
-                        ->heading('Regio & status informatie')
-                        ->description('Gegevens omtrent de regio en status van het woord')
-                        ->icon(Heroicon::OutlinedGlobeEuropeAfrica)
-                        ->iconColor('primary')
-                        ->schema([
-                            Select::make('regions')
-                                ->columnSpanFull()
-                                ->label("Regio's")
-                                ->translateLabel()
-                                ->multiple()
-                                ->relationship(titleAttribute: 'name')
-                                ->optionsLimit(4)
-                                ->preload()
-                                ->minItems(1)
-                                ->required(),
-                            Radio::make('status')
-                                ->columnSpanFull()
-                                ->options(LanguageStatus::class),
-                        ]),
-
-                    Section::make('source-information')
-                        ->compact()
-                        ->heading('Bron gegevens')
-                        ->icon(Heroicon::OutlinedBookOpen)
-                        ->iconColor('primary')
-                        ->description('Registratie formulier voor alle geraadpleegde naslagwerken tijdens het opstellen van het artikel')
-                        ->schema([
-                            Repeater::make('sources')
-                                ->relationship()
-                                ->compact()
-                                ->table([
-                                    Repeater\TableColumn::make('bron')->width(400),
-                                    Repeater\TableColumn::make('referentie')
-                                ])
-                                ->schema([
-                                    Select::make('source_id')
-                                        ->label('bron')
-                                        ->options(User::query()->pluck('name', 'id'))
-                                        ->required()
-                                        ->distinct()
-                                        ->searchable()
-                                        ->disableOptionsWhenSelectedInSiblingRepeaterItems(),
-                                    Textarea::make('referentie')->rows(1)
-                                        ->required()
-                                ])
-                                ->addActionLabel('Naslagwerk toevoegen')
-                                ->defaultItems(1)
-                                ->hiddenLabel()
-                                ->required()
-                        ])
-                ])->columnSpan(9),
-
-            Section::make()
-                ->heading('Redactionele informatie')
-                ->icon(Heroicon::OutlinedUserGroup)
-                ->iconColor('primary')
-                ->compact()
-                ->columns(12)
-                ->schema([
-                    TextEntry::make('author.name')
-                        ->label('Suggestie door')
-                        ->icon(Heroicon::OutlinedUserCircle)
-                        ->iconColor('primary')
-                        ->columnSpanFull(),
-                    TextEntry::make('editor.name')
-                        ->label('Redacteur')
-                        ->icon(Heroicon::OutlinedUserCircle)
-                        ->iconColor('primary')
-                        ->columnSpanFull(),
-                    TextEntry::make('publisher.name')
-                        ->label('Eindredacteut')
-                        ->icon(Heroicon::OutlinedUserCircle)
-                        ->iconColor('primary')
-                        ->placeholder('-')
-                        ->columnSpanFull(),
-                    TextEntry::make('created_at')
-                        ->label('Ingevoerd op')
-                        ->icon(Heroicon::OutlinedCalendar)
-                        ->iconColor('primary')
-                        ->state(fn (Article $record): ?string => $record->created_at?->format('d/m/Y'))
-                        ->columnSpan(6),
-
-                    TextEntry::make('updated_at')
-                        ->icon(Heroicon::OutlinedClock)
-                        ->iconColor('primary')
-                        ->label('Laatst bewerkt')
-                        ->state(fn (Article $record): ?string => $record->updated_at?->diffForHumans())
-                        ->columnSpan(6),
-                ])
-                ->columnSpan(3)
-        ])->columns(12);
-    }
-
-    /**
-     * @return array<int, array<string>>
-     */
-    private static function getToolbarOptions(): array
-    {
-        return [
-            ['bold', 'italic', 'strike', 'link'],
-            ['heading'],
-            ['blockquote', 'bulletList', 'orderedList'],
-            ['table'],
-            ['undo', 'redo'],
-        ];
-    }
-
-    /**
-     * @return array<int, string>|string
-     */
-    private static function getArticleStateOptions(): array|string
-    {
-        if (auth()->user()->user_type->in(enums: [UserTypes::Administrators, UserTypes::Developer])) {
-            return ArticleStates::class;
-        }
-
-        return [
-            ArticleStates::New->value => ArticleStates::New->getLabel(),
-            ArticleStates::Draft->value => ArticleStates::Draft->getLabel(),
-        ];
+        return ArticleForm::configure($schema);
     }
 
     /**
